Move fee decision generation trigger
