+++ service/src/test/kotlin/fi/espoo/evaka/invoicing/service/FeeDecisionGeneratorTest.kt
@@ -61,7 +61,6 @@ class FeeDecisionGeneratorTest {
             txManager,
             dataSource,
             objectMapper,
-            asyncJobRunner,
             env
         )
     }
